ChannelLogger plugin: new noHashPrefix config
The new config defaults to False, so that the plugin' out-of-the-box
behavior does not change.

When noHashPrefix = True, the channel name in the log directories and
files will not be prefixed with a '#'.

This is useful when further processing of the logfiles is performed,
e.g. generating html pages from them due to the hash being used as a
fragment identifier [1] in the URI (so to access the page, the '#' needs
to be escaped to '%23').

[1] http://tools.ietf.org/html/rfc3986#section-3.5